requirements.txt
requests
pytelegrambotapi
python-decouple
black
aiogram
data